+++ Src/LinearSolvers/MLMG/AMReX_MLEBABecLap_K.H
@@ -6,13 +6,6 @@
 
 #include <AMReX_EBCellFlag.H>
 
-namespace amrex {
-    AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
-    Real get_dx_eb (Real kappa) noexcept {
-        return amrex::max(Real(0.3),(kappa*kappa-Real(0.25))/(Real(2.0)*kappa));
-    }
-}
-
 #if (AMREX_SPACEDIM == 2)
 #include <AMReX_MLEBABecLap_2D_K.H>
 #else
