+++ Src/LinearSolvers/MLMG/AMReX_MLNodeTensorLaplacian.cpp
@@ -1,5 +1,5 @@
 #include <AMReX_MLNodeTensorLaplacian.H>
-#include <AMReX_MLNodeLap_K.H>
+#include <AMReX_MLNodeLinOp_K.H>
 #include <AMReX_MLNodeTensorLap_K.H>
 #include <AMReX_MultiFabUtil.H>
