+++ Tests/LinearSolvers/Nodal_Projection_EB/CMakeLists.txt
@@ -1,4 +1,4 @@
-if ( (NOT AMReX_EB) OR NOT (3 IN_LIST AMReX_SPACEDIM))
+if ( (NOT AMReX_EB) OR (NOT AMReX_LINEAR_SOLVERS_INCFLO) OR NOT (3 IN_LIST AMReX_SPACEDIM))
    return()
 endif ()
