+++ Tools/CMake/AMReXConfig.cmake.in
@@ -74,6 +74,8 @@ set(AMReX_AMRLEVEL_FOUND            @AMReX_AMRLEVEL@)
 set(AMReX_EB_FOUND                  @AMReX_EB@)
 set(AMReX_FINTERFACES_FOUND         @AMReX_FORTRAN_INTERFACES@)
 set(AMReX_LSOLVERS_FOUND            @AMReX_LINEAR_SOLVERS@)
+set(AMReX_LSOLVERS_INCFLO_FOUND     @AMReX_LINEAR_SOLVERS_INCFLO@)
+set(AMReX_LSOLVERS_EM_FOUND         @AMReX_LINEAR_SOLVERS_EM@)
 set(AMReX_AMRDATA_FOUND             @AMReX_AMRDATA@)
 set(AMReX_PARTICLES_FOUND           @AMReX_PARTICLES@)
 set(AMReX_P@AMReX_PARTICLES_PRECISION@_FOUND ON)
@@ -129,6 +131,8 @@ set(AMReX_AMRLEVEL                  @AMReX_AMRLEVEL@)
 set(AMReX_EB                        @AMReX_EB@)
 set(AMReX_FINTERFACES               @AMReX_FORTRAN_INTERFACES@)
 set(AMReX_LSOLVERS                  @AMReX_LINEAR_SOLVERS@)
+set(AMReX_LSOLVERS_INCFLO           @AMReX_LINEAR_SOLVERS_INCFLO@)
+set(AMReX_LSOLVERS_EM               @AMReX_LINEAR_SOLVERS_EM@)
 set(AMReX_AMRDATA                   @AMReX_AMRDATA@)
 set(AMReX_PARTICLES                 @AMReX_PARTICLES@)
 set(AMReX_PARTICLES_PRECISION       @AMReX_PARTICLES_PRECISION@)
